refactor(visual): split remaining scaffold renderers
- Move particle-field, oscillation, function-plot, and timeline renderer bodies into a focused core renderer module.

- Keep the main scaffold module as spec extraction, shell helpers, and registry wrappers.

- Extend wrapper regression coverage across all primitive renderers and update cleanup docs.
